[feat]: splitting our code
- In this we splitted code in the individual slices.
- So we used auth and counter slice for the respective functionality
- And that we used into the index file.
