+++ redux-app/src/components/Counter.js
@@ -1,6 +1,7 @@
 import { useDispatch, useSelector } from 'react-redux';
 import classes from './Counter.module.css';
-import { counterActions } from '../store';
+import { counterActions } from '../store/counter.slice';
+
 // import { Component } from 'react';
 
 const Counter = () => {
