+++ redux-app/src/components/Header.js
@@ -1,6 +1,6 @@
 import { useDispatch, useSelector } from 'react-redux';
 import classes from './Header.module.css';
-import { authActions } from '../store';
+import { authActions } from '../store/auth.slice';
 
 const Header = () => {
   // get state for isAuthenticated
