+++ redux-app/src/store/index.js
@@ -1,50 +1,7 @@
 // import {createStore} from 'redux';
-import { createSlice, configureStore } from '@reduxjs/toolkit';
-
-// initial state for counter
-const initialCounterState= {
-  counter: 0,
-  showCounter: true
-}
-
-// create slice for counter
-const counterSlice= createSlice({
-  name: 'counter',
-  initialState: initialCounterState,
-  reducers: {
-    increment(state) {
-      state.counter++;
-    },
-    decrement(state) {
-      state.counter--;
-    },
-    increase(state, action) {
-      state.counter = state.counter + action.payload; 
-    },
-    toggleCounter(state) {
-      state.showCounter = !state.showCounter;
-    },
-  }
-});
-
-// initialstate for authentication
-const initialAuthState = {
-  isAuthenticated: false
-}
-
-// create slice for authentication
-const authSlice= createSlice({
-  name: 'authentication',
-  initialState: initialAuthState,
-  reducers: {
-    login(state) {
-      state.isAuthenticated = true;
-    },
-    logout(state) {
-      state.isAuthenticated = false;
-    }
-  }
-})
+import { configureStore } from '@reduxjs/toolkit';
+import counterReducer from './counter.slice';
+import authReducer from './auth.slice';
 
 // counter reducer
 // const counterReducer = (
@@ -88,15 +45,14 @@ const authSlice= createSlice({
 //   counterSlice.reducer
 // );
 
+
+
 // store with configureStore
 const store = configureStore({
   reducer: {
-    counter: counterSlice.reducer,
-    auth: authSlice.reducer
+    counter: counterReducer,
+    auth: authReducer
   }
 });
-
-export const counterActions= counterSlice.actions;
-export const authActions= authSlice.actions;
 
 export default store;
